+++ examples/spacex/components/DatalayerProvider.tsx
@@ -7,18 +7,25 @@ import {
   fetchExchange,
   createClient,
   persistedExchange,
+  debugExchange,
 } from '@/fuse/client'
 import React from 'react'
 
 export const DatalayerProvider = (props: any) => {
   const [client, ssr] = React.useMemo(() => {
-    const ssr = ssrExchange();
+    const ssr = ssrExchange()
     const { client } = createClient({
       url:
         process.env.NODE_ENV === 'production'
           ? 'https://spacex-fuse.vercel.app/api/fuse'
           : 'http://localhost:3000/api/fuse',
-      exchanges: [cacheExchange, ssr, persistedExchange, fetchExchange],
+      exchanges: [
+        cacheExchange,
+        ssr,
+        persistedExchange,
+        debugExchange,
+        fetchExchange,
+      ],
     })
 
     return [client, ssr]
