+++ packages/core/src/next/pages.ts
@@ -20,6 +20,7 @@ export * from 'urql'
 export const persistedExchange = urqlPersistedExchange({
   enforcePersistedQueries: process.env.NODE_ENV === 'production',
   enableForMutation: true,
+  preferGetForPersistedQueries: true,
   generateHash: (_, document) =>
     Promise.resolve((document as any)['__meta__']['hash']),
 })
